Split block.py 3/8: extract CodeBlock → models/code_block.py
Part 3/8. code_block.py carries the CURRENT CodeBlock (incl. self-heal/escalation: _attempt_self_heal,
_fail_escalation_task, _finalize_recovery_block, _self_heal_enabled, _is_healable_page_failure,
_match_step_for_failing_line) re-extracted from current main — the original #6974 branch was cut from
an older base and would have regressed 6 methods. Imports Block from block_base (cycle-free);
BaseTaskBlock/TASK_TO_BLOCK_STATUS imported lazily. Test patch targets repointed to code_block.
